♻️  refactor: 상수화
