+++ js/product/const/index.js
@@ -1,0 +1,11 @@
+export const RANGE = {
+  MAX_PRICE: 500_000,
+  MIN_PRICE: 100,
+  MAX_COUNT: 5_000,
+  MIN_COUNT: 1,
+  PRICE_UNIT: 10,
+};
+
+export const MESSAGE = {
+  PLZ_CHECK_PRICE_UNIT: `금액은 ${RANGE.PRICE_UNIT}원 단위로 입력하세요.`,
+};
