+++ js/product/helper/product.js
@@ -1,12 +1,6 @@
 import { now } from '../../utils/date.js';
 import { InputValidationError } from '../../utils/validation.js';
-
-export const RANGE = {
-  MAX_PRICE: 500_000,
-  MIN_PRICE: 100,
-  MAX_COUNT: 5_000,
-  MIN_COUNT: 1,
-};
+import { MESSAGE, RANGE } from '../const/index.js';
 
 export const addProduct = (newProduct, prevProducts) =>
   sortByRegisterTime(mergeProducts(newProduct, prevProducts));
@@ -31,13 +25,7 @@ const mergeProducts = (newProduct, prevProducts) => {
 const sortByRegisterTime = (products) =>
   products.sort(({ registerTime: a }, { registerTime: b }) => b - a);
 
-const validateProductInfo = ({ name, count, price }) => {
-  if (!name) InputValidationError.of('name', '상품명을 입력하세요.');
-
-  if (!count) InputValidationError.of('count', '수량을 입력하세요.');
-
-  if (!price) InputValidationError.of('price', '금액을 입력하세요.');
-
-  if (price % 10)
-    InputValidationError.of('price', '금액은 10원 단위로 입력하세요.');
+const validateProductInfo = ({ price }) => {
+  if (price % RANGE.PRICE_UNIT)
+    InputValidationError.of('price', MESSAGE.PLZ_CHECK_PRICE_UNIT);
 };
